anti-cheat-dashboard/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-socketio==5.8.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.15
pydantic==1.10.7
python-dotenv==1.0.0 